FBI Heavy SWAT to MFR Unit
Because, when these guys spawn on Mayhem & Death Wish with the GenSec City Elite, they are still identified as being with the FBI. The most likely reason for this is that back when the Death Wish update hit, the MFRs didn't have a GenSec counterpart. Patches later they finally got one.